% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{add_handlers_back}
\alias{add_handlers_back}
\title{Add a list of functions to an existing HandlerList.}
\usage{
add_handlers_back(existing, ...)
}
\description{
Add a list of functions to an existing HandlerList.
}
