% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_core.R
\name{add_host_exec_env_user_agent_handler}
\alias{add_host_exec_env_user_agent_handler}
\title{Add the execution environment to the User-Agent request header.
TODO: Implement.}
\usage{
add_host_exec_env_user_agent_handler(request)
}
\description{
Add the execution environment to the User-Agent request header.
TODO: Implement.
}
