% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{as.list.struct}
\alias{as.list.struct}
\title{Create a list from an struct object}
\usage{
\method{as.list}{struct}(x, ...)
}
\arguments{
\item{x}{An struct object.}

\item{...}{Other arguments, which will be ignored.}
}
\description{
Create a list from an struct object
}
