% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client_config}
\alias{client_config}
\title{client_config returns a ClientConfig configured for the service.}
\usage{
client_config(service_name, endpoints)
}
\description{
client_config returns a ClientConfig configured for the service.
}
