% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ClientInfo}
\alias{ClientInfo}
\title{A ClientInfo object stores immutable data about a service.}
\usage{
ClientInfo(service_name = "", service_id = "", api_version = "",
  endpoint = "", signing_name = "", signing_region = "",
  json_version = "", target_prefix = "")
}
\description{
A ClientInfo object stores immutable data about a service.
}
