% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{Config}
\alias{Config}
\title{A Config object provides service configuration for service clients.}
\usage{
Config(credentials = structure(list(creds = list(access_key_id = "",
  secret_access_key = "", session_token = "", provider_name = ""), force_refresh
  = TRUE, provider = list(function ()  {     access_key_id <-
  Sys.getenv("AWS_ACCESS_KEY_ID")     secret_access_key <-
  Sys.getenv("AWS_SECRET_ACCESS_KEY")     if (access_key_id != "" &&
  secret_access_key != "") {         creds <- list(access_key_id =
  access_key_id, secret_access_key = secret_access_key, session_token = "",
  provider_name = "")     }     else {         creds <- NULL     }    
  return(creds) }, function ()  {     access_key_id <-
  get_os_env_variable("AWS_ACCESS_KEY_ID")     secret_access_key <-
  get_os_env_variable("AWS_SECRET_ACCESS_KEY")     if (access_key_id != "" &&
  secret_access_key != "") {         creds <- list(access_key_id =
  access_key_id, secret_access_key = secret_access_key, session_token = "",
  provider_name = "")     }     else {         creds <- NULL     }    
  return(creds) }, function ()  {     credentials_path <-
  file.path(get_aws_path(), "credentials")     if
  (!file.exists(credentials_path))          return(NULL)     aws_profile <-
  get_profile_name()     credentials <- ini::read.ini(credentials_path)     if
  (is.null(credentials[[aws_profile]]))          return(NULL)     access_key_id
  <- credentials[[aws_profile]]$aws_access_key_id     secret_access_key <-
  credentials[[aws_profile]]$aws_secret_access_key     if
  (is.null(access_key_id) || is.null(secret_access_key))          return(NULL)  
    if (access_key_id != "" && secret_access_key != "") {         creds <-
  list(access_key_id = access_key_id, secret_access_key = secret_access_key,
  session_token = "", provider_name = "")     }     else {         creds <- NULL
      }     return(creds) }, function ()  {     iam_role <- get_iam_role()    
  if (is.null(iam_role))          return(NULL)     credentials_url <-
  file.path("iam/security-credentials", iam_role)     credentials_response <-
  get_instance_metadata(credentials_url)     if (is.null(credentials_response)) 
          return(NULL)     credentials_response_body <-
  jsonlite::fromJSON(credentials_response$body)     access_key_id <-
  credentials_response_body$AccessKeyId     secret_access_key <-
  credentials_response_body$SecretAccessKey     session_token <-
  credentials_response_body$Token     if (is.null(access_key_id) ||
  is.null(secret_access_key) || is.null(session_token))          return(NULL)   
   if (access_key_id != "" && secret_access_key != "" && session_token != "") { 
         creds <- list(access_key_id = access_key_id, secret_access_key =
  secret_access_key, session_token = session_token, provider_name = "")     }   
   else {         creds <- NULL     }     return(creds) }, function ()  {    
  stop("No credentials provided") })), class = "struct"), endpoint = "",
  endpoint_resolver = "", enforce_should_retry_check = FALSE, region = "",
  disable_ssl = FALSE, max_retries = -1, retryer = NULL,
  disable_param_validation = FALSE, disable_compute_checksums = FALSE,
  s3_force_path_style = FALSE, s3_disable_100_continue = FALSE,
  s3_use_accelerate = FALSE, s3_disable_content_md5_validation = FALSE,
  ec2_metadata_disable_timeout_override = FALSE, use_dual_stack = FALSE,
  sleep_delay = NULL, disable_rest_protocol_uri_cleaning = FALSE)
}
\description{
A Config object provides service configuration for service clients.
}
