% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_ec2query.R
\name{ec2query_unmarshal}
\alias{ec2query_unmarshal}
\title{Unmarshal the response from an EC2 protocol response.}
\usage{
ec2query_unmarshal(request)
}
\description{
Unmarshal the response from an EC2 protocol response.
}
