% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_ec2query.R
\name{ec2query_unmarshal_error}
\alias{ec2query_unmarshal_error}
\title{Unmarshal errors from an EC2 protocol response.}
\usage{
ec2query_unmarshal_error(request)
}
\description{
Unmarshal errors from an EC2 protocol response.
}
