% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_ec2query.R
\name{ec2query_unmarshal_meta}
\alias{ec2query_unmarshal_meta}
\title{Unmarshal metadata from an EC2 protocol response.
Do nothing; the EC2 protocol does not require unmarshalling metadata.}
\usage{
ec2query_unmarshal_meta(request)
}
\description{
Unmarshal metadata from an EC2 protocol response.
Do nothing; the EC2 protocol does not require unmarshalling metadata.
}
