% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_presigned_url.R
\name{generate_presigned_url}
\alias{generate_presigned_url}
\title{Generate a presigned url given a client, its method, and arguments}
\usage{
generate_presigned_url(
  client_method,
  params = list(),
  expires_in = 3600,
  http_method = NULL
)
}
\arguments{
\item{client_method}{(character_: The client method to presign for}

\item{params}{(list): The parameters normally passed to \code{client_method}.}

\item{expires_in:}{The number of seconds the presigned url is valid
for. By default it expires in an hour (3600 seconds)}

\item{http_param}{The http method to use on the generated url. By default,
the http method is whatever is used in the method's model.}
}
\value{
The presigned url character
}
\description{
Generate a presigned url given a client, its method, and arguments
}
