% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_presign_url.R
\name{generate_signed_url}
\alias{generate_signed_url}
\title{Generate a presigned url given a client, its method, and arguments}
\usage{
generate_signed_url(client_method, params = list(), expires_in = 3600)
}
\arguments{
\item{client_method}{(character_: The client method to presign for}

\item{params}{(list): The parameters normally passed to \code{client_method}.}

\item{expires_in:}{The number of seconds the presigned url is valid
for. By default it expires in an hour (3600 seconds)}
}
\value{
The presigned url character
}
\description{
Generate a presigned url given a client, its method, and arguments
}
