% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{HandlerList}
\alias{HandlerList}
\title{Create a list of request handlers.}
\usage{
HandlerList(..., after_each_fn = NULL)
}
\description{
A handler list is a list of functions to apply to a request at a given stage
of a service request (e.g. validate, build, etc.).
}
