% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{Handlers}
\alias{Handlers}
\title{Create the list of request handlers for each stage of a request.}
\usage{
Handlers(validate = list(list = list(), after_each_fn = NULL),
  build = list(list = list(), after_each_fn = NULL), sign = list(list =
  list(), after_each_fn = NULL), send = list(list = list(), after_each_fn =
  NULL), validate_response = list(list = list(), after_each_fn = NULL),
  unmarshal = list(list = list(), after_each_fn = NULL),
  unmarshal_meta = list(list = list(), after_each_fn = NULL),
  unmarshal_error = list(list = list(), after_each_fn = NULL),
  retry = list(list = list(), after_each_fn = NULL), after_retry = list(list
  = list(), after_each_fn = NULL), complete = list(list = list(),
  after_each_fn = NULL))
}
\description{
A handlers object stores the handler lists for each stage of handling a
service request.
}
