% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_empty}
\alias{is_empty}
\title{Check whether an object is empty}
\usage{
is_empty(x)
}
\arguments{
\item{x}{An object.}
}
\description{
Check whether an object is empty, e.g. has no sub-elements or is NA or the
empty string.
}
