% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_empty_xml}
\alias{is_empty_xml}
\title{Check whether an object is empty for xml builds}
\usage{
is_empty_xml(x)
}
\arguments{
\item{x}{An object.}
}
\description{
Check whether an object is empty, e.g. has no sub-elements, is NA
}
\examples{
is_empty_xml(NA) # TRUE
is_empty_xml(list()) # TRUE
is_empty_xml(list(list())) # TRUE

is_empty_xml(1) # FALSE
is_empty_xml("") # FALSE
is_empty_xml(list(1)) # FALSE
is_empty_xml(list(list(1))) # FALSE

}
