% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_jsonrpc.R
\name{jsonrpc_unmarshal_error}
\alias{jsonrpc_unmarshal_error}
\title{Unmarshal errors from a JSON RPC response.}
\usage{
jsonrpc_unmarshal_error(request)
}
\description{
Unmarshal errors from a JSON RPC response.
}
