% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_jsonrpc.R
\name{jsonrpc_unmarshal_meta}
\alias{jsonrpc_unmarshal_meta}
\title{Unmarshal metadata from a JSON RPC response.}
\usage{
jsonrpc_unmarshal_meta(request)
}
\description{
Unmarshal metadata from a JSON RPC response.
}
