% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{list_paginators}
\alias{list_paginators}
\title{List of paginators from paws client.}
\usage{
list_paginators(svc)
}
\arguments{
\item{svc}{paws client to check}
}
\value{
character vector of functions that can be paginated.
}
\description{
List of paginators from paws client.
}
\examples{
\dontrun{
# Note: where svc is a paws client.
list_paginators(svc)
}
}
