% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{locate_credentials}
\alias{locate_credentials}
\title{Locate AWS credentials}
\usage{
locate_credentials(profile = "", anonymous = FALSE)
}
\arguments{
\item{profile}{The name of a profile to use. If not given, then the default profile is used.}

\item{anonymous}{Set anonymous credentials.}
}
\value{
list containing AWS credentials
\itemize{
\item access_key_id - (character) AWS access key ID
\item secret_access_key - (character) AWS secret access key
\item session_token - (character) AWS temporary session token
\item access_token - (character) A token that gives a user permission to access certain resources
\item expiration - (numeric) Indicates the Unix time when an access token will expire.
\item region - (character) The AWS Region used in instantiating the client.
}
}
\description{
Locate AWS credentials
}
