% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{merge_config}
\alias{merge_config}
\title{Merges config lists for paws services}
\usage{
merge_config(orig_cfg, flat_cfg)
}
\arguments{
\item{orig_cfg}{Original config list}

\item{flat_cfg}{Flatten config list}
}
\description{
Allows config list to be flatten from shorthand.
}
\keyword{internal}
