% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{new_request}
\alias{new_request}
\title{Returns a new request for the service API operation and parameters.}
\usage{
new_request(client, operation, params, data)
}
\description{
Returns a new request for the service API operation and parameters.
}
