% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{new_service}
\alias{new_service}
\title{Return an AWS API service object}
\usage{
new_service(metadata, handlers)
}
\arguments{
\item{metadata}{A named list of API metadata. See details.}

\item{handlers}{A set of handlers, e.g. from \code{new_handlers}.}
}
\description{
Return an API service object with information and handlers needed to make
API requests.
}
\details{
The \code{metadata} parameter requires a list like the following:\preformatted{list(
  service_name = "string",
  endpoints = list("region" = "endpoint"),
  service_id = "string",
  api_version = "string",
  signing_name = "string"|NULL,
  json_version = "string",
  target_prefix = "string"
)
}
}
\examples{
# Metadata for the S3 API.
metadata <- list(
  service_name = "s3",
  endpoints = list("us-east-1" = "s3.amazonaws.com"),
  service_id = "S3",
  api_version = "2006-03-01",
  signing_name = NULL,
  json_version = "",
  target_prefix = ""
)

# Handlers for S3.
handlers <- new_handlers("restxml", "v4")

# Build a service object for S3, containing the information necessary to
# build, send, and receive requests.
service <- new_service(metadata, handlers)

}
\seealso{
Other API request functions: \code{\link{new_handlers}},
  \code{\link{new_operation}}, \code{\link{new_request}},
  \code{\link{send_request}}
}
\concept{API request functions}
