% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{Operation}
\alias{Operation}
\title{An operation is the service API operation to be made.}
\usage{
Operation(name = "", http_method = "", http_path = "",
  paginator = list(), before_presign_fn = function ()  { })
}
\description{
An operation is the service API operation to be made.
}
