% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{parse_config}
\alias{parse_config}
\title{Validate and return a service configuration object.}
\usage{
parse_config(config)
}
\arguments{
\item{config}{A list of optional configuration settings.}
}
\description{
Validate and return a service configuration object.
}
\details{
The optional configuration settings can include the following:\preformatted{list(
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string"
  ),
  endpoint = "string",
  region = "string"
)
}
}
\examples{
# Create a config object with custom credentials and endpoint.
config <- parse_config(
  list(
    credentials = list(
      creds = list(
        access_key_id = "abc",
        secret_access_key = "123"
      )
    ),
    endpoint = "https://foo.com"
  )
)

}
