% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{paws_reset_cache}
\alias{paws_reset_cache}
\title{Clear down paws cache environments}
\usage{
paws_reset_cache()
}
\description{
Clears down the cache environments.
\itemize{
\item \code{ini_cache}: an environment that stores the results from \code{read_ini}, mainly used for storing AWS config files.
\item \code{os_env_cache}: an environment that stores environmental variables from Unix Operating Systems.
\item \code{bearer_token_cache}: an environment that stores bearer tokens
}
}
