% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_query.R
\name{query_unmarshal}
\alias{query_unmarshal}
\title{Unmarshal the response from a Query protocol response.}
\usage{
query_unmarshal(request)
}
\description{
Unmarshal the response from a Query protocol response.
}
