% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_query.R
\name{query_unmarshal_error}
\alias{query_unmarshal_error}
\title{Unmarshal errors from a Query protocol response.}
\usage{
query_unmarshal_error(request)
}
\description{
Unmarshal errors from a Query protocol response.
}
