% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_query.R
\name{query_unmarshal_meta}
\alias{query_unmarshal_meta}
\title{Unmarshal metadata from a Query protocol response.}
\usage{
query_unmarshal_meta(request)
}
\description{
Unmarshal metadata from a Query protocol response.
}
