% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_restjson.R
\name{restjson_build}
\alias{restjson_build}
\title{Build the request body for the REST JSON protocol.}
\usage{
restjson_build(request)
}
\description{
Build the request body for the REST JSON protocol.
}
