% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_restjson.R
\name{restjson_unmarshal}
\alias{restjson_unmarshal}
\title{Unmarshal a REST JSON protocol API response.}
\usage{
restjson_unmarshal(request)
}
\description{
Unmarshal a REST JSON protocol API response.
}
