% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_restjson.R
\name{restjson_unmarshal_error}
\alias{restjson_unmarshal_error}
\title{Unmarshal errors from a REST JSON protocol API response.}
\usage{
restjson_unmarshal_error(request)
}
\description{
Unmarshal errors from a REST JSON protocol API response.
}
