% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_restjson.R
\name{restjson_unmarshal_meta}
\alias{restjson_unmarshal_meta}
\title{Unmarshal metadata from a REST JSON protocol API response.}
\usage{
restjson_unmarshal_meta(request)
}
\description{
Unmarshal metadata from a REST JSON protocol API response.
}
