% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_restxml.R
\name{restxml_unmarshal_error}
\alias{restxml_unmarshal_error}
\title{Unmarshal errors from a REST XML response.}
\usage{
restxml_unmarshal_error(request)
}
\description{
Unmarshal errors from a REST XML response.
}
