% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_restxml.R
\name{restxml_unmarshal_meta}
\alias{restxml_unmarshal_meta}
\title{Unmarshal metadata from a REST XML response.}
\usage{
restxml_unmarshal_meta(request)
}
\description{
Unmarshal metadata from a REST XML response.
}
