% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_core.R
\name{sdk_version_user_agent_handler}
\alias{sdk_version_user_agent_handler}
\title{Add the name/version to the User-Agent request header.}
\usage{
sdk_version_user_agent_handler(request)
}
\description{
Add the name/version to the User-Agent request header.
}
