% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{send_request}
\alias{send_request}
\title{Send a request and handle the response}
\usage{
send_request(request)
}
\arguments{
\item{request}{A request, e.g. from \code{new_request}.}
}
\description{
Send a request and handle the response, returning an output object
for the API operation that has been populated with the data received in the
response.
}
