% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tag_add}
\alias{tag_add}
\title{Add a tag to an object.}
\usage{
tag_add(object, tags)
}
\description{
Add a tag to an object, which we can access later using e.g.
`tag_get(object, "locationName")`. This is used to store metadata about an
API shape, such as its location or type.
}
