% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tag_get}
\alias{tag_get}
\title{Get a tag from an object, and return "" if the tag is not present.}
\usage{
tag_get(object, tag)
}
\description{
Get a tag from an object, and return "" if the tag is not present.
}
