% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_v4.R
\name{v4_sign_request_handler}
\alias{v4_sign_request_handler}
\title{Signs an SDK request with the V4 signature.
See https://docs.aws.amazon.com/general/latest/gr/signature-version-4.html.}
\usage{
v4_sign_request_handler(request)
}
\description{
Signs an SDK request with the V4 signature.
See https://docs.aws.amazon.com/general/latest/gr/signature-version-4.html.
}
