% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_core.R
\name{validate_endpoint_handler}
\alias{validate_endpoint_handler}
\title{Check whether the AWS region and service endpoint are valid.}
\usage{
validate_endpoint_handler(request)
}
\description{
Check whether the AWS region and service endpoint are valid.
}
