% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_core.R
\name{validate_parameters_handler}
\alias{validate_parameters_handler}
\title{Check whether the parameters are valid prior to being sent.
TODO: Implement.}
\usage{
validate_parameters_handler(request)
}
\description{
Check whether the parameters are valid prior to being sent.
TODO: Implement.
}
