% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_core.R
\name{validate_req_sig_handler}
\alias{validate_req_sig_handler}
\title{Ensure that the request's signature doesn't expire before it is sent.}
\usage{
validate_req_sig_handler(request)
}
\description{
Ensure that the request's signature doesn't expire before it is sent.
}
