% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers_core.R
\name{validate_response_handler}
\alias{validate_response_handler}
\title{Check whether the service's response was ok.}
\usage{
validate_response_handler(request)
}
\description{
Check whether the service's response was ok.
}
