% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{handlers}
\alias{handlers}
\alias{handlers_set}
\alias{handlers_add_back}
\alias{handlers_add_front}
\title{Set or add to handler lists}
\usage{
handlers_set(...)

handlers_add_back(existing, ...)

handlers_add_front(existing, ...)
}
\arguments{
\item{...}{One or more request-handling functions to add.}

\item{existing}{An existing request handler list.}
}
\description{
Set or add to a list of request-handling functions. Request-handling
functions must take a request and return a request.
}
