% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{locate_credentials}
\alias{locate_credentials}
\title{Locate AWS credentials}
\usage{
locate_credentials(profile = "", anonymous = FALSE)
}
\arguments{
\item{profile}{The name of a profile to use. If not given, then the default profile is used.}

\item{anonymous}{Set anonymous credentials.}
}
\value{
list containing AWS credentials
}
\description{
Locate AWS credentials
}
