% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{new_handlers}
\alias{new_handlers}
\title{Return request handlers for a service}
\usage{
new_handlers(protocol, signer)
}
\arguments{
\item{protocol}{Protocol: \code{ec2query}, \code{jsonrpc}, \code{query}, \code{rest}, \code{restjson},
or \code{restxml}.}

\item{signer}{Signer: \code{v2} or \code{v4}.}
}
\description{
Return request handlers for a given protocol and request signer.
}
\examples{
# Get the handlers needed for an API using REST-JSON and AWS signature V4.
handlers <- new_handlers("restjson", "v4")

}
\seealso{
Other API request functions: 
\code{\link{new_operation}()},
\code{\link{new_request}()},
\code{\link{new_service}()},
\code{\link{send_request}()}
}
\concept{API request functions}
