% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{new_operation}
\alias{new_operation}
\title{Return an API operation object}
\usage{
new_operation(name, http_method, http_path, paginator,
  before_presign_fn = NULL)
}
\arguments{
\item{name}{The API operation name.}

\item{http_method}{The HTTP method, e.g. \code{"GET"} or \code{"POST"}.}

\item{http_path}{The HTTP path.}

\item{paginator}{Currently unusued.}

\item{before_presign_fn}{Currently unused.}
}
\description{
Return an API operation object
}
