% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{new_service}
\alias{new_service}
\title{Return an AWS API service object}
\usage{
new_service(metadata, handlers)
}
\arguments{
\item{metadata}{A named list of API metadata. See details.}

\item{handlers}{A set of handlers, e.g. from \code{new_handlers}.}
}
\description{
Return an API service client used to make requests.
}
\details{
The \code{metadata} parameter requires a list like the following:\preformatted{list(
  service_name = "string",
  endpoints = list("region" = "endpoint"),
  service_id = "string",
  api_version = "string",
  signing_name = "string"|NULL,
  json_version = "string",
  target_prefix = "string"
)
}
}
