% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{send_request}
\alias{send_request}
\title{Send a request to the service.
TODO: Retry the request in case of error.}
\usage{
send_request(request)
}
\description{
Send a request to the service.
TODO: Retry the request in case of error.
}
