% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{send_request}
\alias{send_request}
\title{Send a request and handle the response}
\usage{
send_request(request)
}
\arguments{
\item{request}{A request, e.g. from \code{new_request}.}
}
\description{
Send a request and handle the response. Build the HTTP request, send it to
AWS, interpret the response, and throw an error if the response is not ok.
}
\examples{
# Send a request and handle the response.
\dontrun{resp <- send_request(req)}

}
\seealso{
Other API request functions: 
\code{\link{new_handlers}()},
\code{\link{new_operation}()},
\code{\link{new_request}()},
\code{\link{new_service}()}
}
\concept{API request functions}
